split sidebar into components
