+++ apps/postgres-new/components/sidebar/sidebar-footer/theme-dropdown.tsx
@@ -1,13 +1,13 @@
 'use client'
 
 import { Moon, Sun } from 'lucide-react'
-import { Button } from './ui/button'
+import { Button } from '~/components/ui/button'
 import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuTrigger,
-} from './ui/dropdown-menu'
+} from '~/components/ui/dropdown-menu'
 import { useTheme } from 'next-themes'
 import { cn } from '~/lib/utils'
